+++ mkmc/KmersSamplesStruct.h
@@ -10,5 +10,5 @@ struct KmersSamplesStruct
 	static const unsigned SIZE = _SIZE;
 
 	CKmer<_SIZE> minKmer;
-	const std::vector<size_t>& kMersCounts;
+	const std::vector<uint64_t>& kMersCounts;
 };
